# fix(cli): dedup `preceding_auxiliary_symbols`