@clason 2024-08-26
I just did a full crater test on nvim-treesitter (333 languages, 1455 query files), and this indeed made a significant difference (<7min instead of 19min) with a flat memory profile never peaking above 6 Gb. In particular, the two outliers C# and QMLJS are gone now. All query tests are still passing so there seems to be no regression in the generated parsers.